Moved ray_tracer functions in their corresponding files
