+++ Makefile
@@ -42,7 +42,9 @@ SCENE_OBJECTS_SRCS =	ambient_light.c				\
 						scene_object_utils.c
 SCENE_OBJECTS_OBJS =	${addprefix ${OBJ_DIR}/, ${notdir ${SCENE_OBJECTS_SRCS:.c=.o}}}
 
-RAY_TRACER_SRCS =	ray_init.c
+RAY_TRACER_SRCS =	ray_init.c					\
+					compute_intersection.c		\
+					compute_light.c
 RAY_TRACER_OBJS =	${addprefix ${OBJ_DIR}/, ${notdir ${RAY_TRACER_SRCS:.c=.o}}}
 
 MLX_SRCS =	mlx.c
