+++ includes/ray_tracer.h
@@ -6,7 +6,7 @@
 /*   By: lgreau <lgreau@student.42heilbronn.de>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/22 14:19:51 by lgreau            #+#    #+#             */
-/*   Updated: 2024/05/24 13:09:39 by lgreau           ###   ########.fr       */
+/*   Updated: 2024/05/24 14:06:43 by lgreau           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 # define RAY_TRACER_H
 
 void		init_ray(void);
-uint32_t	intersect_ray(t_vector3 *ray);
+uint32_t	compute_intersection(t_vector3 *ray);
 uint32_t	compute_light(double intersect, t_vector3 *ray, t_scene_object *obj);
 
 #endif
