+++ src/ray_tracer/compute_intersection.c
@@ -1,0 +1,49 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   compute_intersection.c                             :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: lgreau <lgreau@student.42heilbronn.de>     +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/05/24 14:06:23 by lgreau            #+#    #+#             */
+/*   Updated: 2024/05/24 14:07:40 by lgreau           ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "miniRT.h"
+
+/**
+ * @brief Computes if there's an intersection with any solid object in the scene
+ * & returns the computed light
+ *
+ * @param ray
+ * @return uint32_t
+ */
+uint32_t	compute_intersection(t_vector3 *ray)
+{
+	t_program	*program;
+	int			index;
+	int			min;
+	double		min_value;
+	double		intersect;
+
+	program = get_program();
+	index = -1;
+	min = index;
+	min_value = INFINITY;
+	while (++index < program->object_count)
+	{
+		if ((int) program->objects[index].type == SPHERE)
+		{
+			intersect = get_obj_intersect()[program->objects[index].type](ray, &program->objects[index]);
+			if (intersect < min_value)
+			{
+				min = index;
+				min_value = intersect;
+			}
+		}
+	}
+	if (min_value < INFINITY)
+		return (compute_light(min_value, ray, &program->objects[min]));
+	return (color_scal_mult(get_object(AMBIENT_LIGHT)->s_ambient_light.color, get_object(AMBIENT_LIGHT)->s_ambient_light.intensity));
+}
